Mike | changes based on comments in PR, mostly moving things around and fixing errors
