+++ components/Player/Pages/playerHeroesColumns.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { transformations, getPercentWin } from '../../../utility';
-import { TablePercent } from '../../Table';
+import { TablePercent } from '../../Visualizations';
 
 export const playerHeroesOverviewColumns = [{
   displayName: 'Hero',
