+++ components/Table/index.js
@@ -1,7 +1,5 @@
 export { default } from './Table';
 export { default as createTable, createTables } from './createTable';
 export { default as TableContainer } from './TableContainer';
-export { default as TablePercent } from './TablePercent';
 export { default as TableHeroImage } from './TableHeroImage';
 export { default as TableLink } from './TableLink';
-export { default as TableKDA } from './TableKDA';
