+++ components/Visualizations/TablePercent.css
