+++ components/Visualizations/index.js
@@ -1,0 +1,2 @@
+export { default as TablePercent } from './TablePercent';
+export { default as KDA } from './KDA';
